# needrelay-frontend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: NeedRelay API
  version: 0.1.0
  description: Crisis and emergency resource coordination API
servers:
  - url: /
...
paths:
...
  /api/admin/organizers:
    get:
      tags: [admin]
      operationId: listOrganizers
      summary: List organizers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Organizers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizerResponse'
  /api/admin/organizers/{organizerId}/ban:
    post:
      tags: [admin]
      operationId: banOrganizer
      summary: Ban organizer
      security:
        - bearerAuth: []
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Banned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerResponse'
  /api/admin/organizers/{organizerId}/unban:
    post:
      tags: [admin]
      operationId: unbanOrganizer
      summary: Unban organizer
      security:
        - bearerAuth: []
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Unbanned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerResponse'
  /api/admin/organizers/{organizerId}/role:
    patch:
      tags: [admin]
      operationId: adminUpdateOrganizerRole
      summary: Update organization member role
      security:
        - bearerAuth: []
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerResponse'
...
  /api/me/organization:
    get:
      tags: [organization]
      operationId: getMyOrganization
      summary: Get my organization
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
    put:
      tags: [organization]
      operationId: updateMyOrganization
      summary: Update my organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
  /api/me/organization/members:
    get:
      tags: [organization]
      operationId: listOrganizationMembers
      summary: List organization members
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationMemberResponse'
  /api/me/organization/members/{organizerId}:
    patch:
      tags: [organization]
      operationId: updateMemberRole
      summary: Update member role
      security:
        - bearerAuth: []
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMemberResponse'
    delete:
      tags: [organization]
      operationId: removeMember
      summary: Remove member
      security:
        - bearerAuth: []
      parameters:
        - name: organizerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Removed
  /api/invites:
    get:
      tags: [invites]
      operationId: listInvites
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InviteResponse'
    post:
      tags: [invites]
      operationId: createInvite
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteResponse'
  /api/invites/{inviteId}:
    delete:
      tags: [invites]
      operationId: revokeInvite
      summary: Revoke invite
      security:
        - bearerAuth: []
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    OrganizerResponse:
      type: object
      required: [id, email, displayName, role, active, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [ADMIN, ORGANIZER] }
        active: { type: boolean }
        organizationId: { type: string, format: uuid, nullable: true }
        organizationName: { type: string, nullable: true }
        organizationRole:
          allOf: [{ $ref: '#/components/schemas/OrganizationRole' }]
          nullable: true
        createdAt: { type: string, format: date-time }
...
    OrganizationResponse:
      type: object
      required: [id, name, active, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    UpdateOrganizationRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
...
    OrganizationMemberResponse:
      type: object
      required: [id, email, displayName, organizationRole, active, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        displayName: { type: string }
        organizationRole:
          $ref: '#/components/schemas/OrganizationRole'
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
    UpdateMemberRoleRequest:
      type: object
      required: [organizationRole]
      properties:
        organizationRole:
          $ref: '#/components/schemas/OrganizationRole'
...
    CreateInviteRequest:
      type: object
      properties:
        email: { type: string, format: email, nullable: true }
        daysValid: { type: integer, minimum: 1, maximum: 90, default: 7 }
        organizationId: { type: string, format: uuid, nullable: true }
        organizationRole:
          allOf: [{ $ref: '#/components/schemas/OrganizationRole' }]
          nullable: true
    InviteResponse:
      type: object
      required: [id, token, expiresAt, createdAt, organizationRole]
      properties:
        id: { type: string, format: uuid }
        token: { type: string }
        email: { type: string, nullable: true }
        organizationId: { type: string, format: uuid, nullable: true }
        organizationName: { type: string, nullable: true }
        organizationRole:
          $ref: '#/components/schemas/OrganizationRole'
        expiresAt: { type: string, format: date-time }
        usedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        emailSentAt: { type: string, format: date-time, nullable: true }
        emailError: { type: string, nullable: true }
...
    OrganizationRole:
      type: string
      enum: [ADMIN, USER]
...