# needrelay-frontend/src/api/generated/models/organizationMemberResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { OrganizationRole } from './organizationRole';

export interface OrganizationMemberResponse {
  id: string;
  email: string;
  displayName: string;
  organizationRole: OrganizationRole;
  active: boolean;
  createdAt: string;
}