# needrelay-frontend/src/api/generated/models/updateOrganizationRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrganizationRequest {
  name: string;
  /** @nullable */
  description?: string | null;
}